requirements.txt
accelerate
aiohttp
aiosignal
attrs
certifi
charset-normalizer
contourpy
cycler
datasets
dill
filelock
fonttools
frozenlist
fsspec
gurobipy
huggingface-hub
idna
jinja2
joblib
kiwisolver
markupsafe
matplotlib
mpmath
multidict
multiprocess
networkx
numpy
packaging
pandas
patsy
pillow
psutil
pulp
pyarrow
pyarrow-hotfix
pyparsing
python-dateutil
pytz
pyyaml
regex
requests
safetensors
scikit-learn
scipy
seaborn
six
statsmodels
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing-extensions
tzdata
urllib3
xxhash
yarl
pyJoules
pynvml
